fix(package): types not exported correctly
